moved groupOptions to its own folder
