+++ src/views/ModelPlan/ReadOnly/_components/FilterView/Modal/index.tsx
@@ -10,6 +10,8 @@ import {
   Link
 } from '@trussworks/react-uswds';
 
+import groupOptions from '../util';
+
 type FilterViewModalProps = {
   closeModal: () => void;
 };
@@ -19,27 +21,6 @@ const FilterViewModal = ({ closeModal }: FilterViewModalProps) => {
   const history = useHistory();
 
   const [filteredGroup, setFilteredGroup] = useState('');
-
-  const groupOptions = [
-    { value: 'ccw', label: 'Chronic Conditions Warehouse (CCW)' },
-    { value: 'cmmi', label: 'CMMI Cost Estimate' },
-    {
-      value: 'cbosc',
-      label: 'Consolidated Business Operations Support Center (CBOSC)'
-    },
-    {
-      value: 'dfsdm',
-      label: 'Division of Financial Services and Debt Management (DFSDM)'
-    },
-    { value: 'ipc', label: 'Innovation Payment Contractor (IPC)' },
-    {
-      value: 'iddoc',
-      label: 'Innovative Design, Development, and Operations Contract (IDDOC)'
-    },
-    { value: 'mdm', label: 'Master Data Management (MDM)' },
-    { value: 'oact', label: 'Office of the Actuary (OACT)' },
-    { value: 'pbg', label: 'Provider Billing Group (PBG)' }
-  ];
 
   const handleSubmit = (value: string) => {
     if (value === 'view-all') {
