+++ src/views/ModelPlan/ReadOnly/_components/FilterView/util.ts
@@ -1,0 +1,22 @@
+const groupOptions = [
+  { value: 'ccw', label: 'Chronic Conditions Warehouse (CCW)' },
+  { value: 'cmmi', label: 'CMMI Cost Estimate' },
+  {
+    value: 'cbosc',
+    label: 'Consolidated Business Operations Support Center (CBOSC)'
+  },
+  {
+    value: 'dfsdm',
+    label: 'Division of Financial Services and Debt Management (DFSDM)'
+  },
+  { value: 'ipc', label: 'Innovation Payment Contractor (IPC)' },
+  {
+    value: 'iddoc',
+    label: 'Innovative Design, Development, and Operations Contract (IDDOC)'
+  },
+  { value: 'mdm', label: 'Master Data Management (MDM)' },
+  { value: 'oact', label: 'Office of the Actuary (OACT)' },
+  { value: 'pbg', label: 'Provider Billing Group (PBG)' }
+];
+
+export default groupOptions;
